# Conflicts:
#	src/main/java/com/lukgru/galgo/runner/reproduction/SimpleReproductionRunner.java
#	src/main/java/com/lukgru/galgo/runner/solution/SimpleSolutionSeeker.java
#	src/test/java/com/lukgru/galgo/runner/reproduction/SimpleReproductionRunnerTest.java

+++ src/main/java/com/lukgru/galgo/runner/fitness/SimpleFitnessCalculator.java
@@ -18,7 +18,7 @@ public class SimpleFitnessCalculator<T> implements FitnessCalculator<T> {
     public void compute(Population<T> population) {
         population.getIndividuals().forEach(individual -> {
             T value = individual.getValue();
-            Integer fitnessValue = fitnessFunction.getFitnessFunction().apply(value);
+            Double fitnessValue = fitnessFunction.getFitnessFunction().apply(value);
             individual.setFitnessScore(fitnessValue);
         });
     }
